parameter for cont script

+++ features/dcc-helpers/install-helpers.sh
@@ -10,5 +10,8 @@ fi
 envsubst '$DCC_VERSION,$DCC_BINARY,$DCC_REPL' < disclaimer.sh > /disclaimer.sh
 chmod +x /disclaimer.sh
 
-envsubst '$DCC_VERSION,$DCC_BINARY,$DCC_REPL' < cont.sh > /usr/local/bin/cont
-chmod +x /usr/local/bin/cont
+if [ -f "cont.sh" ]
+then
+    envsubst '$DCC_VERSION,$DCC_BINARY,$DCC_REPL' < cont.sh > /usr/local/bin/cont
+    chmod +x /usr/local/bin/cont
+fi
